requirements.txt
numpy>=1.21.0
setuptools>=50.0.0
wheel>=0.36.0
pandas>=1.3.0
requests>=2.25.1
python-dotenv>=0.19.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
openai>=1.0.0
tiktoken>=0.5.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.1
langgraph>=0.1.0
scikit-learn==0.24.0
faiss-cpu>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
newspaper3k>=0.2.8
feedparser>=6.0.0
streamlit>=1.28.0
chromadb>=0.4.0